move Navbar component into it's own file
